chore: refactor token code to isolate each variant
