+++ formats/Tokeniser_Factory.py
@@ -1,5 +1,7 @@
-from formats.Coco_Tokens import CoCoToken, RsDosToken
-from formats.Dragon_Tokens import DragonToken, DragonDosToken
+from formats.Coco_Tokens import CoCoToken
+from formats.Coco_Rsdos_Tokens import RsDosToken
+from formats.Dragon_Tokens import DragonToken
+from formats.Dragon_Dos_Tokens import DragonDosToken
 
 
 def find_tokeniser(options):
